requirements.txt
requests~=2.31.0
psycopg2-binary~=2.9.9
pandas~=2.2.1
matplotlib~=3.8.3

#Native
#os~=
#re~=
#csv~=
#pathlib~=
